requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
langsmith==0.1.85
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
